requirements.txt
certifi==2020.6.20
cffi==1.14.3
cryptography==3.1.1
decorator==4.4.2
pycparser==2.20
python-dotenv==0.14.0
python-telegram-bot==12.8
six==1.15.0
tornado==6.0.4
